remove excess from builder, more organizational changes, token() -> build()
